requirements.txt
onnxruntime-gpu==1.14.1
numpy
opencv-python
python-utils
icecream
pandas
ipython
scikit-learn
tensorflow==1.12.0